Updated to `0.4.0` and fix publishing issue
There were some changes in `ConsoleMessageSource` enum and thus, its a
breaking change, we have to bump the major version.

I made that enum `non_exhaustive` for future changes.

The issue with publishing was that it was using an older version, maybe
because we are not updating the version of the `sys` crate normally and
instead just modifying the hash revision.

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>