Updated `ul-next-sys` to implement `libloading` (build fails)
Partial update to the `libloading` feature, added the support for it
in the `sys` crate, this works well. But the main crate is failing to build because of the changes introduced here.

We changed how we generate the bindings, used a small crate `generate` to do that instead of using `bindgen` binary in a shell script.

What we do, is generate `load`/`link` bindings for `ultralight` libs and `appcore` lib, and we can allow the user to choose whether to include `appcore` or not during `loading` or `linking`

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>