Updated `ul-next` to use the new `Library` for everything
This new struct, `Library` holds the ultralight library.
Its an abstraction that covers `linked` and `loaded` versions, so which version you choose to use, the rest of the code will be the same.

The issue with this now is that `ultralight` by design (since we use the C API) relies on global variables and things are done in the form of `platform::instance().do_something` i.e. singletons.

With the introduction of this variable (library) that must be used at all times, it makes using the library a bit more complicated. We need to think of a better way to do these stuff.

Examples are still failing, will be fixed next.

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>